Using new_euler1d instead of euelr1d app
