+++ cpp/modmesh/view/RManager.cpp
@@ -211,7 +211,6 @@ void RManager::setUpMenu()
         this->addApplication(QString("euler1d"));
         this->addApplication(QString("linear_wave"));
         this->addApplication(QString("bad_euler1d"));
-        this->addApplication(QString("new_euler1d"));
     }
 
     {
